Create a new page for similarity search testing and refactor votestars function
This commit introduces a new page at `public_html/stuff/tag-zeroclip-vote.php` for collecting scores to test similarity search capabilities. The page fetches a random 'subject' tag from the database, displays it with voting stars, and loads thumbnails for user assessment. It also includes a "Next" button and voting guidance.

This commit also refactors the `smarty_function_votestars` function by moving it to a central location in `libs/geograph/functions.inc.php` and registering it globally. This removes code duplication and improves maintainability.
