+++ libs/geograph/functions.inc.php
@@ -1573,3 +1573,18 @@ function https_request_with_ip_override(string $url, string $override_ip, array 
     return file_get_contents($url, false, $context);
 }
 
+function smarty_function_votestars($params) {
+    global $CONF;
+    static $last;
+
+    $type = $params['type'];
+    $id = $params['id'];
+    $names = array('','Hmm','Below average','So So','Good','Excellent');
+    foreach (range(1,5) as $i) {
+        print "<a href=\"javascript:void(record_vote('$type',$id,$i));\" title=\"{$names[$i]}\"><img src=\"{$CONF['STATIC_HOST']}/img/star-light.png\" width=\"14\" height=\"14\" alt=\"$i\" onmouseover=\"star_hover($id,$i,5)\" onmouseout=\"star_out($id,5)\" name=\"star$i$id\"/></a>";
+    }
+    if ($last != $type) {
+        print " (<a href=\"/help/voting\">about</a>)";
+    }
+    $last = $type;
+}
