+++ public_html/admin/moderation.php
@@ -527,8 +527,6 @@ $images->assignSmarty($smarty, 'unmoderated');
 $style = $USER->getStyle();
 $smarty->assign('maincontentclass', 'content_photo'.$style);
 
-    $smarty->register_function("votestars", "smarty_function_votestars");
-
 if (!empty($_GET['full']))
 	$smarty->assign('full',1);
 
@@ -539,19 +537,4 @@ $smarty->display('admin_moderation.tpl',$style);
 
 // This tries to force the page to reload, when user presses back, mimicking the old behaviour of Cache-Control:no-store which no longer works for BFCache
 enforceNoStoreBFCache();
-
-
-
-function smarty_function_votestars($params) {
-    global $CONF;
-    static $last;
-
-    $type = $params['type'];
-    $id = $params['id'];
-    $names = array('','Hmm','Below average','So So','Good','Excellent');
-    foreach (range(1,5) as $i) {
-        print "<a href=\"javascript:void(record_vote('$type',$id,$i));\" title=\"{$names[$i]}\"><img src=\"{$CONF['STATIC_HOST']}/img/star-light.png\" width=\"14\" height=\"14\" alt=\"$i\" onmouseover=\"star_hover($id,$i,5)\" onmouseout=\"star_out($id,5)\" name=\"star$i$id\"/></a>"; 
-    }
-    $last = $type;
-}
