+++ public_html/editimage.php
@@ -756,27 +756,8 @@ else
 
 
 
-    $smarty->register_function("votestars", "smarty_function_votestars");
-
 $smarty->display($template, $cacheid);
 
 // This tries to force the page to reload, when user presses back, mimicking the old behaviour of Cache-Control:no-store which no longer works for BFCache
 enforceNoStoreBFCache();
-
-
-function smarty_function_votestars($params) {
-    global $CONF;
-    static $last;
-
-    $type = $params['type'];
-    $id = $params['id'];
-    $names = array('','Hmm','Below average','So So','Good','Excellent');
-    foreach (range(1,5) as $i) {
-        print "<a href=\"javascript:void(record_vote('$type',$id,$i));\" title=\"{$names[$i]}\"><img src=\"{$CONF['STATIC_HOST']}/img/star-light.png\" width=\"14\" height=\"14\" alt=\"$i\" onmouseover=\"star_hover($id,$i,5)\" onmouseout=\"star_out($id,5)\" name=\"star$i$id\"/></a>";
-    }
-    if ($last != $type) {
-        print " (<a href=\"/help/voting\">about</a>)";
-    }
-    $last = $type;
-}
