+++ public_html/stuff/votes-rater.php
@@ -133,7 +133,7 @@ print "<!-- $query -->\n";
 		}
 		print "</a>";
 		print "<div id=\"votediv{$image->gridimage_id}\">";
-		smarty_function_votestars(array('type'=>$type,'id'=>$row['id']));
+		$smarty->display('string:{votestars type=\''.$type.'\' id='.$row['id'].'}');
 		if (!empty($votes[$row['id']]))
 			print " (existing: {$votes[$row['id']]})</div></div>";
 		else
@@ -148,18 +148,3 @@ print "<!-- $query -->\n";
 
 
 
-function smarty_function_votestars($params) {
-	global $CONF;
-	static $last;
-	
-	$type = $params['type'];
-	$id = $params['id'];
-	$names = array('','Hmm','Below average','So So','Good','Excellent');
-	foreach (range(1,5) as $i) {
-		print "<a href=\"javascript:void(record_vote('$type',$id,$i));\" title=\"{$names[$i]}\"><img src=\"{$CONF['STATIC_HOST']}/img/star-light.png\" width=\"14\" height=\"14\" alt=\"$i\" onmouseover=\"star_hover($id,$i,5)\" onmouseout=\"star_out($id,5)\" name=\"star$i$id\"/></a>";
-	}
-	if ($last != $type) {
-		print " (<a href=\"/help/voting\">about</a>)";
-	} 
-	$last = $type;
-}
